FB - Frontend size and type validation for files (#47)
* Button and table styling

* Filebrowser frontend blocks non-text filetypes and files over 1MB

* Some placeholder window.alert and window.confirm, can change them to
  less pop-uppy versions later
* Made things into tables
* Also played around with css
* The theme is kinda close to Tokyo Night... Could go all the way.

* Utilized startsWith()

* introduces a githook for pre-push

* Fixed a small typo in github actions

+++ frontend/src/dashboard/FileList.tsx
@@ -9,11 +9,11 @@ function FileList({
 	fileList: UserFile[] | null;
 	refreshFileList: () => void;
 }): JSX.Element {
-	if (!fileList) {
-		return <p>Loading really slow...</p>;
-	}
+	if (!fileList) return <p>Loading really slow...</p>;
+	if (fileList.length === 0) return <p>You lead a fileless existence.</p>;
 
 	async function handleDelete(id: string) {
+		if (!window.confirm("Are you sure you want to delete this file?")) return;
 		try {
 			const result = await fetch(`/api/files/${id}`, {
 				method: "DELETE",
